:whale: Replace shell process rather than spawning a child process
exec replaces the shell process, so that uwsgi will run as PID 1. The
situation before this ran uwsgi as child processes of the shell
process (which had PID 1). When stopping containers, the Docker daemon
sends a TERM signal to PID 1, and after a grace time of 30s (default)
sends the KILL signal. By default, the bash process does not propagate
signals to the child processes. Making uwsgi PID 1 ensures these
signals are respected properly.